ledger_app/requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
bcrypt==5.0.0
PyJWT==2.12.1
python-dotenv==1.2.2